% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_p2.R
\name{dt_p2}
\alias{dt_p2}
\title{creates dataset for two binomial proportions}
\usage{
dt_p2(n, pc, pt, add_x = FALSE, xs_y0 = 0.6, xs_y1 = 0.2, xw = 0.6,
  add_xcont = FALSE)
}
\arguments{
\item{n}{integer, number of observation per group, 1:1 ratio assumed}

\item{pc}{numeric, probability of event in group 'c'}

\item{pt}{numeric, probability of event in group 't'}

\item{add_x}{logic, Default: FALSE, specifies whether to add binarry
variable x}

\item{xs_y0}{numeric, Default: 0.6, for strong correlation between x and y,
define conditional probability of x given y=0}

\item{xs_y1}{numeric, Default: 0.2, for strong correlation between x and y,
define conditional probability of x given y=1}

\item{xw}{numeric, Default: 0.6, for weak correlation between x and y,
define prrobability of x (irrespective of the value of y)}

\item{add_xcont}{logic, Default: FALSE, specifies whether to add continuous
variable x}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
creates dataset for two binomial proportions
}
\details{
DETAILS
}
\examples{
dt_p2(100, 0.7, 0.65, add_x = TRUE)
}
\seealso{
\code{\link[stats]{Binomial}}
 \code{\link[purrr]{map}}
}
