% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_bin.R
\name{gibbs_bin}
\alias{gibbs_bin}
\title{Gibbs sampling for beta-binomial distribution}
\usage{
gibbs_bin(dt, B = 1000, y.m = "y.m")
}
\arguments{
\item{dt}{tibble}

\item{B}{numeric, Default: 1000, number of iterations}

\item{y.m}{string, Default: 'y.m', column with incomplete binary data}
}
\value{
complete dataset after B iterations of Gibbs sampler
}
\description{
performs Gibbs sampling for beta-binomial distribution
}
\seealso{
\code{\link[stats]{Beta}},\code{\link[stats]{Binomial}}
}
