% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb_wn.R
\name{lb_wn}
\alias{lb_wn}
\title{lower bound wilson}
\usage{
lb_wn(z, phat, n_obs)
}
\arguments{
\item{z}{numeric, quantile of either standard normal or t distribution
coresponding to the desired confidence level 1- alpha}

\item{phat}{numeric, proporrtion estimate}

\item{n_obs}{numeric, number of observations}
}
\value{
numeric
}
\description{
calculates lower bound of (1-alpha)100\% confidence interval
using Wilson's method
}
\examples{
lb_wn(1.96, 0.6, 100)
}
