% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb_wn_ign.R
\name{lb_wn_ign}
\alias{lb_wn_ign}
\title{lower bound wilson, ignorable mi}
\usage{
lb_wn_ign(z, qhat, n_obs, rn)
}
\arguments{
\item{z}{numeric, quantile of t distribution coresponding to the desired
confidence level 1- alpha}

\item{qhat}{numeric}

\item{n_obs}{integer, number of observations}

\item{rn}{interger}
}
\value{
numeric
}
\description{
calculates lower bound of (1-alpha)100\% confidence interval
using Wilson's method following MI assuming ignorability
}
\examples{
lb_wn_ign(1.96, 0.8, 100, 0.7)
}
