% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb_wn_p2.R
\name{lb_wn_p2}
\alias{lb_wn_p2}
\title{lower bound newcombe}
\usage{
lb_wn_p2(pc, lb_pc, ub_pc, pt, lb_pt, ub_pt)
}
\arguments{
\item{pc}{numeric, probability in group 'c'}

\item{lb_pc}{numeric, lower bound of probability in group 'c', Wilson's
method}

\item{ub_pc}{numeric, upper bound of probability in group 'c', Wilson's
method}

\item{pt}{numeric, probability in group 't'}

\item{lb_pt}{numeric, lower bound of probability in group 't', Wilson's
method}

\item{ub_pt}{numeric, upper bound of probability in group 't', Wilson's
method}
}
\value{
numeric
}
\description{
calculates lower bound of (1-alpha)100\% confidence interval
using Newcombe's method
}
\examples{
lb1 <- lb_wn(1.96, 0.6, 100)
ub1 <- ub_wn(1.96, 0.6, 100)
lb2 <- lb_wn(1.96, 0.8, 100)
ub2 <- ub_wn(1.96, 0.8, 100)
lb_wn_p2(0.6, lb1, ub1, 0.8, lb2, ub2)
}
