% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi.R
\name{mi}
\alias{mi}
\title{multitple imputation for proportions}
\usage{
mi(dt, n_mi, m_mi = 0, mu_k = 1, sd_k = 0, ym, phat_out = TRUE)
}
\arguments{
\item{dt}{dataframe/tibble}

\item{n_mi}{positive integer, number of multiply imputed subject level values}

\item{m_mi}{non-negative integer, Default: 0, number of muttiply imputed
values for multitplier k}

\item{mu_k}{numeric, Default: 1, mean value for normal distribution of
multiplier k}

\item{sd_k}{numeric, Default: 0, standard deviation value for normal
distribution of multiplier k}

\item{ym}{integer, incomplete binary column}

\item{phat_out}{logic, Default TRUE}
}
\value{
tibble, summary per imputation or if phat_out=FALSE, returns imputed
 data (available only for one-stage imputation)
}
\description{
employs one/two-stage multiple imputation for proportions
}
\details{
DETAILS
}
\examples{
dt <- tibble::tibble(y = rbinom(100,1,0.6))
dt$ym <- c(rep(NA, 10), dt$y[11:100])
 mi(dt, 5, ym = 'ym')
 mi(dt, 2, 10, mu_k = 1.3, sd_k = 0.1, ym = 'ym')
}
\seealso{
\code{\link[bin2mi]{sf_count}}
 \code{\link[tibble]{tibble}}
 \code{\link[purrr]{map}}
 \code{\link[rlang]{sym}}
 \code{\link[tidyr]{unnest}}
}
