% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_x0.R
\name{mp_x0}
\alias{mp_x0}
\title{conditional missing probability for missing at random}
\usage{
mp_x0(do_tar, p_y1, mp_x1, xs_ass = c("strong", "weak"), xs_y0 = 0.6,
  xs_y1 = 0.2, xw = 0.6)
}
\arguments{
\item{do_tar}{numeric, target dro-out rate}

\item{p_y1}{numeric, probability of y=1}

\item{mp_x1}{numeric, missing probability conditional on x=1}

\item{xs_ass}{character, Default: c("strong", "weak")}

\item{xs_y0}{numeric, Default: 0.6, for strong correlation between x and y,
define conditional probability of x given y=0}

\item{xs_y1}{numeric, Default: 0.2, for strong correlation between x and y,
define conditional probability of x given y=1}

\item{xw}{numeric, Default: 0.6, for weak correlation between x and y,
define prrobability of x (irrespective of the value of y)}
}
\value{
nymeric, conditional probability of missing when x=1
}
\description{
calculates probability of missing conditional on x=0
}
\details{
DETAILS
}
\examples{
mp_x0(do_tar = 0.3, p_y1 = 0.65, mp_x1 = 0.4, xs_ass = 'strong')
}
