% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_y0.R
\name{mp_y0}
\alias{mp_y0}
\title{conditional missing probability for missing not at random}
\usage{
mp_y0(do_tar, mp_y1, p_y1)
}
\arguments{
\item{do_tar}{numeric, target dro-out rate}

\item{mp_y1}{numeric, missing probability conditional on y=1}

\item{p_y1}{numeric, probability of y=1}
}
\value{
numeric
}
\description{
calculates probability of missing conditional on y=0
}
\details{
the aim of this function is to calculate probability of missing
conditional on y=0, in order to impose mnar missingness structure
}
\examples{
mp_y0(0.1, 0.14, 0.65)
}
