% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p1_eval.R
\name{p1_eval}
\alias{p1_eval}
\title{propotion estimation evaluation}
\usage{
p1_eval(dt, trt, ptrt)
}
\arguments{
\item{dt}{tibble which contains simulation results}

\item{trt}{group for which the evaluation is done}

\item{ptrt}{true proportion for group trt}
}
\value{
tibble
}
\description{
calculates coverage, confidence inteval length, proportion of
times confidence interal was outside the limits, proportions of times the
width of the confidence interval was zero
}
\seealso{
\code{\link[purrr]{map}}
 \code{\link[dplyr]{filter}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{summarise}}
}
