% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2_eval.R
\name{p2_eval}
\alias{p2_eval}
\title{difference in propotions evaluation}
\usage{
p2_eval(dt, m2)
}
\arguments{
\item{dt}{tibble which contains simulation results}

\item{m2}{numeric, pre-specified margin}
}
\value{
tibble
}
\description{
calculates coverage, confidence inteval length, proportion of
times confidence interal was outside the limits, proportions of times the
width of the confidence interval was zero
}
\details{
DETAILS
}
\seealso{
\code{\link[purrr]{map}}
}
