% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2_full_ci.R
\name{p2_full_ci}
\alias{p2_full_ci}
\title{condifence interval for difference in binomial proportions}
\usage{
p2_full_ci(dt_mle, method = c("wald", "fm", "wn"), alpha)
}
\arguments{
\item{dt_mle}{tibble containing mle estimated from p2_mle output}

\item{method}{character string indicating one of the following methods for
confidence interval construction: "wald", "fm", "wn".}

\item{alpha}{numeric, alpha level for two-sided confidence interval}
}
\value{
tibble
}
\description{
calculates (1-alpha) 100\% confidence interval for difference
in binomial proportions
}
\details{
DETAILS
}
\examples{
dt <- dt_p2(100, 0.7, 0.65)
dtmle <- p2_mle(dt, m2 = 0.05)
p2_full_ci(dtmle, method = "wald", alpha = 0.05)
}
\seealso{
\code{\link[dplyr]{mutate}},\code{\link[dplyr]{select}}
 \code{\link[stats]{Normal}}
 \code{\link[purrr]{map2}}
}
