% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2_diff_mi.R
\name{p2d_mi}
\alias{p2d_mi}
\title{estimates for difference in binomial proportions}
\usage{
p2d_mi(dt, m2)
}
\arguments{
\item{dt}{tibble, output after mi}

\item{m2}{numeric, pre-specified margin}
}
\value{
tibble
}
\description{
calculates estiamtes for difference in proportions, normal
approximation variance and resticted maximum likelihood variance terms
}
\details{
DETAILS
}
\seealso{
\code{\link[dplyr]{mutate}}
}
