% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_rmle.R
\name{p_rmle}
\alias{p_rmle}
\title{restricted maximum likelihood estimate: proportion difference}
\usage{
p_rmle(m2, nt, nc, pc, pt)
}
\arguments{
\item{m2}{numeric, difference in binomial proportions under the null}

\item{nt}{integer, number of observations in group 't'}

\item{nc}{integer, number rof observation in group 'c'}

\item{pc}{numeric, estimated binomial proportion in group 'c'}

\item{pt}{numeric, estimated binomial proportion in group 't't}
}
\value{
numeric
}
\description{
calculates restricted maximum likelihood estimate under
the null hypothesis of difference in binomial proportions
}
\details{
returns resticted maximum likelihood estimate for binomial
proportion in group C
}
\examples{
p_rmle(0.1, 100, 100, 0.65, 0.57)
}
