% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_count.R
\name{sf_count}
\alias{sf_count}
\title{success/fail counts}
\usage{
sf_count(dt, ym)
}
\arguments{
\item{dt}{dataframe/tibble, includes incomplete binary column}

\item{ym}{integer, incomplete binary column}
}
\value{
tibble
}
\description{
counts number of successes and failures for the incomplete
column ym
}
\examples{
dt <- tibble::tibble(y = rbinom(100,1,0.6))
dt$ym <- c(rep(NA, 10), dt$y[11:100])
sf_count(dt, 'ym')
}
\seealso{
\code{\link[rlang]{sym}}
 \code{\link[tidyr]{spread}}
}
