% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ub_wn.R
\name{ub_wn}
\alias{ub_wn}
\title{upper bound wilson}
\usage{
ub_wn(z, phat, n_obs)
}
\arguments{
\item{z}{numeric, quantile of either standard normal or t distribution
coresponding to the desired confidence level 1- alpha}

\item{phat}{numeric, phat proporrtion estimate}

\item{n_obs}{numeric, n_obs number of observations}
}
\value{
numeric
}
\description{
calculates upper bound of (1-alpha)100\% confidence interval
using Wilson's method
}
\examples{
ub_wn(1.96, 0.6, 100)
}
